Support showAll and includeEmpties also for the thanks page.
